refactor: order judge.py like generate.py, and split path resolution out of targets.py
Two related placement fixes, both from the same review comment.

`vera_cli/judge.py` had `register` first while claiming in its docstring to be
"structured exactly like vera_cli/generate.py", which stopped being true once
#191's reordering arrived. Now it matches: run -> resolve_configs ->
_from_cli/_from_config -> helpers -> _execute -> register. Verified as a pure
move by diffing sorted file content before and after; the only content changes
are the docstring and the handler comment.

`vera_cli/targets.py` had grown a second concern. Its docstring says "discovery
and validation of complete target manifests", but the bottom of the file held
`config_path`, `config_paths`, `config_dir`, and `rubrics_from_config` -- none
of which are about target manifests. They apply one rule, that a config path is
repo-relative and must exist, to whatever field a command hands them, and they
are thin wrappers over `path_from_root`/`existing_file`, which already live in
`config.py`. So they move there.

Two of those four were mine, added earlier today: I moved them toward the right
rule but the wrong module. `targets_from_config` sat exactly on the seam
between the two groups, which is what the comment was pointing at.

Dependency direction is unchanged -- `targets.py` already imported `config.py`
and never the reverse -- so this is a move plus import updates, no cycles.
Both module docstrings and the `vera_cli/README.md` layout table now state the
split.

Suite: 1,060 passed, 8 deselected, coverage 75.49%. Ruff and pyright clean.
