+++ vera_cli/README.md
@@ -15,8 +15,8 @@ not built yet.
 | `../vera.py` | Root parser and dispatcher. Registers each command, routes to its handler, turns `ConfigError` into a standard CLI error. |
 | `generate.py` | The `generate` command: flags, input resolution, and the call into the generation domain. The reference implementation for new commands. |
 | `judge.py` | The `judge` command, same five steps. Differs from legacy `judge.py` in two ways, both recorded in [../docs/vera-cli-use-cases.md](../docs/vera-cli-use-cases.md): no single-conversation mode, and no implicit working-directory output. |
-| `config.py` | Shared input handling: loading config JSON, the config-or-flags rule, path resolution, resolved-run rendering. |
-| `targets.py` | Target discovery and manifest validation — turning a target name into concrete, verified file paths. |
+| `config.py` | Shared input handling: loading config JSON, the config-or-flags rule, turning config-stated strings into verified absolute paths, resolved-run rendering. |
+| `targets.py` | Target discovery and manifest validation — turning a target *name* into concrete, verified file paths. Builds on `config.py`'s path helpers rather than owning them. |
 | `../utils/config_schema.py` | *Not a CLI module.* Shared canonical types (`RunConfig` and friends) in the leaf `utils/` layer, so domain packages may consume them too. Validation and serialization only; no parsing, no defaults. |
 
 ## Adding a command
